cepa/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
bleach==3.1.0
cffi==1.12.3
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
Django==2.2.1
django-admin==1.3.2
django-cors-headers==3.0.2
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.9.4
entrypoints==0.3
Fiona==1.8.6
flake8==3.7.7
GDAL==2.4.1
geopandas==0.5.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflect==2.1.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
munch==2.3.2
mypy==0.711
mypy-extensions==0.4.1
mysqlclient==1.4.2
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyproj==2.2.1
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.1
queuelib==1.5.0
screen==1.0.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
sqlacodegen==2.0.1
SQLAlchemy==1.3.5
sqlparse==0.3.0
svn==0.3.46
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
